tests(e2e): refactor fixtures and make teardowns more robust
