+++ rdmo/management/tests/e2e/conftest.py
@@ -1,20 +1,19 @@
 import pytest
 
-from _pytest.fixtures import SubRequest
-from playwright.sync_api import Page
+from playwright.sync_api import Page, expect
 
-from rdmo.core.tests.e2e.conftest import (  # noqa: F401
+from rdmo.core.tests.e2e.fixtures import (  # noqa: F401
     PLAYWRIGHT_TIMEOUT,
     _set_django_allow_async_unsafe,
+    allow_live_server_host,
     authenticated_context,
     authenticated_page,
     django_db_setup,
-    fail_on_js_error,
 )
 
 
-@pytest.fixture
-def e2e_username(scope="session") -> str:
+@pytest.fixture(scope="session")
+def e2e_username() -> str:
     """Fixture to specify which user should be authenticated.
     This can be overridden in individual test modules or fixtures.
     """
@@ -22,29 +21,24 @@ def e2e_username(scope="session") -> str:
 
 
 @pytest.fixture
-def page_single(django_db_setup, live_server, browser, authenticated_page: Page) -> Page:  # noqa: F811
+def page_single(transactional_db, live_server, authenticated_page: Page) -> Page:  # noqa: F811
     """Navigates the authenticated page to /management."""
     authenticated_page.goto("/management")  # Navigate to the projects section
-    authenticated_page.wait_for_load_state()  # maybe not needed
+    expect(authenticated_page.get_by_role("heading", name="Management")).to_be_visible()
     return authenticated_page
 
 
 @pytest.fixture
-def page_multisite(django_db_setup, live_server, browser, authenticated_page: Page, settings) -> Page:  # noqa: F811
+def page_multisite(transactional_db, live_server, authenticated_page: Page, settings) -> Page:  # noqa: F811
     """Enables the multisite and navigates the authenticated page to /management."""
     settings.MULTISITE = True
     authenticated_page.goto("/management")  # Navigate to the projects section
-    authenticated_page.wait_for_load_state()  # maybe not needed
+    expect(authenticated_page.get_by_role("heading", name="Management")).to_be_visible()
     return authenticated_page
 
 
 @pytest.fixture
-def page(request, django_db_setup, authenticated_page) -> Page:  # noqa: F811
-    """Allows for the specific page fixtures to be parameters and returns a default fixture for page"""
-    try:
-        return request.getfixturevalue(request.param)
-    except AttributeError as e:
-        if isinstance(request, SubRequest) and request.fixturename == "page":
-            # set a default return fixture for "page"
-            return request.getfixturevalue("page_multisite")
-        raise e from e
+def page(request, transactional_db) -> Page:
+    """Allow specific page fixtures to be selected indirectly and default to multisite."""
+    fixture_name = getattr(request, "param", "page_multisite")
+    return request.getfixturevalue(fixture_name)
